// packages/core/phase1/types/ResultClass.ts
enum ResultClass {
  ADJOURNMENT = "Adjournment",
  ADJOURNMENT_WITH_JUDGEMENT = "Adjournment with Judgement",
  JUDGEMENT_WITH_FINAL_RESULT = "Judgement with final result",
  UNRESULTED = "Unresulted",
  SENTENCE = "Sentence",
  ADJOURNMENT_POST_JUDGEMENT = "Adjournment post Judgement",
  ADJOURNMENT_PRE_JUDGEMENT = "Adjournment pre Judgement",
  APPEAL_OUTCOME = "Appeal Outcome"
}

export default ResultClass